Adds change-breaking for transports with limits
Websockets aren't meant for limitless payload sizes, and some services have hard limits on the request size allowed. Batching handled breaking up large sets of changes to be sent to the server in chunks. This update goes a step further to break up changes that are extremely large themselves. It endeavors to do it early when the change is first made in PatchesDoc, but it will handle it later in the batching step of sync if it isn't able to there. While these large changes are not common, they can happen, especially with a large paste operation with *tons* of text or a "duplicate workspace/project" operation.

Another step towards making Patches more resilient in more use-cases.